Ensure the inflight metrics are properly updated when rejecting
Currently, the inflight metrics only update when a successful listener is acquired.
However, in cases where all inflight requests are tied up and everything is being rejected,
the inflight counter never updates. This causes the metric to eventually fall off and "appear"
as 0 inflight. This change always publishes the inflight distribution to ensure it is always
published with its current value.